feat(initial): Initial Version
This is the very first version of my walking script, it's not perfect,
but it works well enough that I can shared it in case other people find
it useful. Please read the instructions in the README.md file for
guidance on how to create `web/walkingPath.jsonl` and the settings you'll
need to put into `config.js`.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>