fix(csv): gracefully handle new columns
The original version of this script died if it encountered a column in
the CSV that it didn't have a mapping for. This change makes it ignore
columns that it doesn't know about. It also adds preliminary parsing
support for two new columns, `Walking Time` and `Total Walking Time`,
which will be used to help record how much time it took me to walk the
distance. This will eventually let me track faster and slower days.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>