feat(actions): build walkingPath.jsonl in GitHub action
This brings in a `Pipfile` to make it so the python environment is
reproducible. It adds a script in the file called `convert` that is used
to automatically create the `.jsonl` file needed for the directions and
path to show up on the final map.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>