feat(verbosity): add --verbose flag to dataprep.py
This was more as a result of moving stuff to GitHub actions to push, but
the dataprep.py script was really chatty. This makes it so it only
outputs the number of lines changed unless the --verbose flag is passed
to the program.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>